git-svn-id: svn+ssh://green.exnet.com/rw/SVN-repositories/public/reutils/trunk@243 7a6dc53a-678a-4e31-bf73-3c095f94da06

+++ javasrc/org/hd/d/edh/FUELINSTUtils.java
@@ -391,7 +391,7 @@ public final class FUELINSTUtils
                     {
                     final float corr = (float) StatUtils.ComputePearsonCorrelation(gridIntensity, fuelMW);
                     // Retain correlation only if sane / finite.
-                    if(!Double.isNaN(corr) && !Double.isInfinite(corr))
+                    if(!Float.isNaN(corr) && !Float.isInfinite(corr))
                         { correlationIntensityToFuel.put(fuel, corr); }
                     }
                 }
